+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.596] - 2026-08-25
+
 - Fix: [Inconsistency in how the Output views follow editor font size](https://github.com/BetterThanTomorrow/calva/issues/3276)
 - Add settings: `calva.outputViews.fontSizeScale`, and `calva.outputViews.wordWrap`
 - [Add API for subscribing to REPL session updates (connect/disconnect, etc)](https://github.com/BetterThanTomorrow/calva/issues/3274)